+++ game/missiongenerator/airsupport.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
+from collections import defaultdict
 
 from dataclasses import dataclass, field
 from datetime import timedelta
 from typing import Optional, TYPE_CHECKING
+from uuid import UUID
 
 if TYPE_CHECKING:
     from game.radio.radios import RadioFrequency
@@ -50,7 +52,22 @@ class JtacInfo:
 
 
 @dataclass
+class LogisticsInfo:
+    """Logistics information."""
+
+    pilot_names: list[str] = field(default_factory=list)
+    pickup_zone: str = field(default_factory=str)
+    drop_off_zone: str = field(default_factory=str)
+    target_zone: str = field(default_factory=str)
+    blue: bool = field(default_factory=bool)
+    preload: bool = field(default_factory=bool)
+
+
+@dataclass
 class AirSupport:
     awacs: list[AwacsInfo] = field(default_factory=list)
     tankers: list[TankerInfo] = field(default_factory=list)
     jtacs: list[JtacInfo] = field(default_factory=list)
+    logistics: dict[UUID, LogisticsInfo] = field(
+        default_factory=lambda: defaultdict(LogisticsInfo)
+    )